Create NIVETHITHA
My project is about displaying the calendar. Here it displays the date,day and month of that particular year.For that we need to know some basic things about the calendar. We would need to know about the leap year.For February month we are checking whether it has 28 or 29 days the code is created such that it displays all the date ,month and day of a year.The dates in the calendar are not always uniform. They keep on changing from year to year.For example in the year 2000 march 21 is in Wednesday and in 2001 it is in Thursday.
First the number of days in a month is given as an input.Then the months are given.The year we want to see is given as an input .Then the February month is checked for the no of days.If 28 it returns true else false.Then the output is displayed.